backend/requirements.txt
arxiv==1.4.2
Flask==2.1.2
Flask_Cors==3.0.10
openai==0.24.0
python-dotenv==0.21.0